Add support for multiple OreDictionary names
When creating modpacks, it's very possible that you'll want to mark an item that already has an existing OreDictionary entry as being an ore (or ingot, etc), while also wanting to keep the original OreDictionary values.
Since OreDictionary does not support reordering or removing values, this makes it incredibly hard for modpack devs to mark something like this with the tools available (without writing java code, that is, which most modpack devs will want to avoid) to COFH mods, since they only check the first OreDictionary entry, and not ones after it.
This tries to fix that by simply checking all of the OreDictionary entries on an item, hopefully making this much easier for modpack devs in future.